loilibreback/requirements.txt
fastapi[standard]
supabase
uvicorn
aisuite[openai, mistral]
python-dotenv
pytest
pytest-asyncio
httpx